perf: split source mention completion
